# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: MeuPaciente API
  version: "1.0.0"
  description: |
    API do sistema de prontuário veterinário para profissionais autônomos,
    com suporte inicial a pacientes cães e gatos.

    ## Autenticação
    Todos os endpoints (exceto `/v1/auth/*`) exigem um JWT Bearer token
    obtido via `POST /v1/auth/login`.

    ## Isolamento de dados
    Todos os recursos são isolados por `user_id` derivado do JWT.
    Um veterinário nunca acessa dados de outro.

    ## Paginação
    Endpoints de coleção usam paginação offset-based com `page` e `per_page`.
    Valores padrão: `page=1`, `per_page=20`. Máximo: `per_page=100`.

    ## Erros
    Todos os erros seguem RFC 7807 (`application/problem+json`).
servers:
  - url: https://api.meupaciente.com.br/v1
    description: Produção
  - url: http://localhost:3000/v1
    description: Desenvolvimento local
...
paths:
...
  /vaccines:
    get:
      tags: [Vaccines]
      summary: Listar vacinas
      operationId: listVaccines
      description: |
        Endpoint principal para o histórico vacinal. Use `patient_id` para
        listar as vacinas de um paciente específico.
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
        - name: patient_id
          in: query
          description: Filtrar por paciente
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lista paginada de vacinas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaccineListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Vaccines]
      summary: Registrar vacina aplicada
      operationId: createVaccine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVaccineRequest"
            example:
              patient_id: "550e8400-e29b-41d4-a716-446655440000"
              appointment_id: "770e8400-e29b-41d4-a716-446655440002"
              name: "V10"
              manufacturer: "Zoetis"
              batch: "L12345"
              application_date: "2024-06-15"
              next_dose_date: "2025-06-15"
              notes: "Sem reações adversas"
      responses:
        "201":
          description: Vacina registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vaccine"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /vaccines/{vaccine_id}:
    parameters:
      - $ref: "#/components/parameters/VaccineId"

    get:
      tags: [Vaccines]
      summary: Obter vacina por ID
      operationId: getVaccine
      responses:
        "200":
          description: Vacina retornada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaccineDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [Vaccines]
      summary: Atualizar vacina
      operationId: updateVaccine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVaccineRequest"
      responses:
        "200":
          description: Vacina atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vaccine"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

    delete:
      tags: [Vaccines]
      summary: Remover vacina
      operationId: deleteVaccine
      responses:
        "204":
          description: Vacina removida
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

# ---------------------------------------------------------------------------
# COMPONENTS
# ---------------------------------------------------------------------------
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Número da página
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Itens por página
...
    VaccineId:
      name: vaccine_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  # ── SCHEMAS ───────────────────────────────────────────────────────────────
  schemas:
...
    Sex:
      type: string
      enum: [MALE, FEMALE, UNKNOWN]
      description: |
        MALE = Macho, FEMALE = Fêmea, UNKNOWN = Não informado
    Patient:
      type: object
      required: [id, tutor_id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        tutor_id:
          type: string
          format: uuid
        name:
          type: string
          example: "Mimi"
        species:
          type: string
          nullable: true
          description: |
            Espécie do paciente. Campo texto livre — os valores iniciais
            suportados são "Cão" e "Gato", mas a API aceita qualquer valor.
          example: "Gato"
        sex:
          $ref: "#/components/schemas/Sex"
        breed:
          type: string
          nullable: true
          example: "Persa"
        color:
          type: string
          nullable: true
          example: "Branca"
        birth_date:
          type: string
          format: date
          nullable: true
          example: "2020-03-15"
        photo_url:
          type: string
          format: uri
          nullable: true
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    Vaccine:
      type: object
      required: [id, patient_id, name, application_date, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        patient_id:
          type: string
          format: uuid
        appointment_id:
          type: string
          format: uuid
          nullable: true
          description: Atendimento em que a vacina foi aplicada, se houver
        name:
          type: string
          example: "V10"
        manufacturer:
          type: string
          nullable: true
          example: "Zoetis"
        batch:
          type: string
          nullable: true
          example: "L12345"
        application_date:
          type: string
          format: date
          example: "2024-06-15"
        next_dose_date:
          type: string
          format: date
          nullable: true
          description: Previsão da próxima dose/reforço
          example: "2025-06-15"
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    VaccineDetail:
      allOf:
        - $ref: "#/components/schemas/Vaccine"
        - type: object
          properties:
            patient:
              $ref: "#/components/schemas/Patient"
    CreateVaccineRequest:
      type: object
      required: [patient_id, name, application_date]
      properties:
        patient_id:
          type: string
          format: uuid
        appointment_id:
          type: string
          format: uuid
          nullable: true
          description: Deve pertencer ao mesmo paciente informado
        name:
          type: string
          minLength: 1
          maxLength: 120
        manufacturer:
          type: string
          nullable: true
          maxLength: 120
        batch:
          type: string
          nullable: true
          maxLength: 60
        application_date:
          type: string
          format: date
        next_dose_date:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
    UpdateVaccineRequest:
      type: object
      properties:
        appointment_id:
          type: string
          format: uuid
          nullable: true
          description: Deve pertencer ao mesmo paciente da vacina
        name:
          type: string
          minLength: 1
          maxLength: 120
        manufacturer:
          type: string
          nullable: true
          maxLength: 120
        batch:
          type: string
          nullable: true
          maxLength: 60
        application_date:
          type: string
          format: date
        next_dose_date:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
    VaccineListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Vaccine"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # -- Shared --------------------------------------------------------------
    Pagination:
      type: object
      required: [page, per_page, total, total_pages]
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 45
        total_pages:
          type: integer
          example: 3
    Problem:
      description: RFC 7807 Problem Details
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: "https://meupaciente.com.br/errors/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: "O campo 'email' deve ser um endereço de e-mail válido."
        instance:
          type: string
          format: uri
          example: "/v1/users"
        errors:
          type: array
          description: Detalhes por campo (para erros de validação)
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Deve ser um e-mail válido"

  # ── RESPONSES ─────────────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Token ausente, inválido ou expirado
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://meupaciente.com.br/errors/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "Token de acesso inválido ou expirado."
...
    NotFound:
      description: Recurso não encontrado
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://meupaciente.com.br/errors/not-found"
            title: "Not Found"
            status: 404
            detail: "O recurso solicitado não foi encontrado."
...
    UnprocessableEntity:
      description: Erro de validação dos dados enviados
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://meupaciente.com.br/errors/validation-error"
            title: "Validation Error"
            status: 422
            detail: "Um ou mais campos estão inválidos."
            errors:
              - field: "email"
                message: "Deve ser um e-mail válido"
              - field: "password"
                message: "Deve ter pelo menos 8 caracteres"
security:
  - BearerAuth: []